Move moderation related code
